move the nix parsing into the library crate
